Run codegen

+++ pkg/querier/testing/mock_querier.go
@@ -394,12 +394,12 @@ func (mr *MockAgentBGPPolicyInfoQuerierMockRecorder) GetBGPPeerStatus(ctx any) *
 }
 
 // GetBGPPolicyInfo mocks base method.
-func (m *MockAgentBGPPolicyInfoQuerier) GetBGPPolicyInfo() (string, string, int32, int32) {
+func (m *MockAgentBGPPolicyInfoQuerier) GetBGPPolicyInfo() (string, string, int64, int32) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetBGPPolicyInfo")
 	ret0, _ := ret[0].(string)
 	ret1, _ := ret[1].(string)
-	ret2, _ := ret[2].(int32)
+	ret2, _ := ret[2].(int64)
 	ret3, _ := ret[3].(int32)
 	return ret0, ret1, ret2, ret3
 }
